Update refresh-dependencies script

+++ docker/destroy-dependencies.sh
@@ -1,5 +1,7 @@
 rm -rf ../vendor
 rm -rf ../node_modules
 rm -rf ../tests/node_modules
-rm ../web/sites/default/settings.local.php
+if [ -f ../web/sites/default/settings.local.php ]; then
+  rm ../web/sites/default/settings.local.php
+fi
